split mir module
