+++ src/lir/mod.rs
@@ -112,4 +112,3 @@ impl Term {
         }
     }
 }
-
