+++ src/mir/lower.rs
@@ -1,57 +1,10 @@
-use std::fmt;
-
 use crate::{
-    ast::*,
+    ast::{BinOp, Block, Literal, Located, Location, Name, Node, UnOp},
+    mir::Term,
     ty::{expect_ty, ty_check, Binding, Ty, TyResult},
-    LangError, LangResult,
 };
 
-#[derive(Debug)]
-pub enum Term<'a> {
-    Var(Name<'a>),
-    Abs(Binding<'a>, Box<Located<Term<'a>>>),
-    UnaryOp(UnOp, Box<Located<Term<'a>>>),
-    BinaryOp(BinOp, Box<Located<Term<'a>>>, Box<Located<Term<'a>>>),
-    App(Box<Located<Term<'a>>>, Box<Located<Term<'a>>>),
-    Lit(Literal),
-    Cond(
-        Box<Located<Term<'a>>>,
-        Box<Located<Term<'a>>>,
-        Box<Located<Term<'a>>>,
-    ),
-    Let(
-        Located<Name<'a>>,
-        Box<Located<Term<'a>>>,
-        Box<Located<Term<'a>>>,
-    ),
-    Seq(Box<Located<Term<'a>>>, Box<Located<Term<'a>>>),
-    Fix(Box<Located<Term<'a>>>),
-}
-
-impl<'a> fmt::Display for Term<'a> {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Term::Var(var) => write!(f, "{}", var),
-            Term::Abs(Binding { name, ty }, term) => write!(f, "(λ{}:{}. {})", name, ty, term),
-            Term::UnaryOp(op, term) => write!(f, "({}{})", op, term),
-            Term::BinaryOp(op, t1, t2) => write!(f, "({} {} {})", t1, op, t2),
-            Term::App(t1, t2) => write!(f, "({} {})", t1, t2),
-            Term::Lit(literal) => write!(f, "{}", literal),
-            Term::Cond(t1, t2, t3) => write!(f, "(if {} then {} else {})", t1, t2, t3),
-            Term::Let(name, t1, t2) => write!(f, "(let {} = {} in {})", name, t1, t2),
-            Term::Seq(t1, t2) => write!(f, "{} ; {}", t1, t2),
-            Term::Fix(t1) => write!(f, "(fix {})", t1),
-        }
-    }
-}
-
-impl<'a> Term<'a> {
-    pub fn from_ast(blk: Located<Block<'a>>) -> LangResult<Located<Self>> {
-        lower_blk(blk).map_err(LangError::Ty)
-    }
-}
-
-fn lower_blk<'a>(blk: Located<Block<'a>>) -> TyResult<Located<Term<'a>>> {
+pub fn lower_blk<'a>(blk: Located<Block<'a>>) -> TyResult<Located<Term<'a>>> {
     let mut terms = blk.content.into_iter().rev().map(lower_node);
     if let Some(term) = terms.next() {
         let mut term = term?;
